requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
gspread==6.1.2
google-auth==2.35.0
google-api-python-client==2.149.0
yookassa==3.3.0
APScheduler==3.10.4
Pillow==10.4.0
moviepy==1.0.3
aiofiles==23.2.1
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
gspread==6.1.2
google-auth==2.35.0
google-api-python-client==2.149.0
yookassa==3.3.0
APScheduler==3.10.4
Pillow==10.4.0
moviepy==1.0.3
aiofiles==23.2.1